rmm: Apply clippy lints
[Command]
$ cargo clippy --lib -p islet_rmm -- \
	-A clippy::comparison_chain \
	-A clippy::crate_in_macro_def \
	-A clippy::empty_loop \
	-A clippy::explicit_auto_deref \
	-A clippy::from_over_into \
	-A clippy::len_without_is_empty \
	-A clippy::manual_range_contains \
	-A clippy::match_like_matches_macro \
	-A clippy::missing_safety_doc \
	-A clippy::new_without_default \
	-A clippy::type_complexity

+++ rmm/src/exception/trap/syndrome.rs
@@ -71,7 +71,7 @@ impl From<u32> for Syndrome {
             }
             0b10_0110 => Syndrome::SPAlignmentFault,
             0b11_1100 => Syndrome::Brk((origin & ESR_EL2::ISS_BRK_CMT as u32) as u16),
-            ec => Syndrome::Other(ec as u32),
+            ec => Syndrome::Other(ec),
         }
     }
 }
